src/requirements.txt
appdirs==1.4.3
attrs==19.3.0
asgiref==3.2.3

certifi==2019.11.28
chardet==3.0.4
Click==7.0

Django==3.0.3

djangorestframework==3.11.0
docutils==0.16

idna==2.9
jmespath==0.9.5
Markdown==3.2.1

python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
toml==0.10.0
